refactor(pwa): colocate expense edit route state
Summary:
- Move the expense edit history marker helper from shared lib into the expense
  detail route folder.
- Update the detail and edit route imports to use the route-local helper.

Rationale:
- The marker is specific to the expense detail/edit route pair and should stay
  near the routes that own the history behavior.

Tests:
- vp run check
- vp exec playwright test e2e/expense-log.spec.ts -g "returns to (the expense log after editing|expense detail when backing out)"
